+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # Addressable 2.8.0
+- fixed encoding ipv6 literals
 - dropping support for Ruby 2.0 and 2.1
 - adding Ruby 3.0 compatibility for development tasks
 - drop support for `rack-mount` and remove related Addressable::Template#generate